Delete corresponding routes when deleting a content item
This change provides a means to delete routes, from the Publishing
API, via the Content Store.

On an architectural level, the Publishing API doesn't directly talk to
the Router API, all communication goes through the Content Store. The
Publishing API can delete content items from the Content Store,
e.g. if you unpublish some content with a vanish type.

This change means that deletions from the Content Store will mean that
the Content Store will delete the relevant routes from the router.

One practical issue which this helps with is removing content/routes
when they are no longer wanted. This is tricky with content with
prefix routes. Say you have two bits of content, A and B where the
routes are:

  /A (type prefix)
  /A/B

If you wanted to remove B, such that A was responsible for /A/B, then
the vanish unpublishing type, with this change to the Content Store
would be a way of doing that.

This change doesn't address the issue of deleting routes when a
ContentItem no longer specifies that route, which I think is another
somewhat related deficiency.