Hard delete routes from the Router
This avoids the Router API service converting these routes to gone
routes, and gets it to actually delete them.

I think this behaviour is better, it avoids state building up in the
Router API and also means that the unpublishing type of vanish will
work, as the routes will be deleted from the router.