Test against Publishing API pacts as part of build
Publishing API publishes a suite of pact tests each time it builds on a
branch and tests these against content-store on the deployed to
production branch. However the inverse of this seems to have been
overlooked and it was possible to merge commits into content-store which
broke these pact tests.