Include redirects in redirect schema output
This is to allow applications other than router the option to know about
which redirects a redirect content item has.

The reason for this is to enable a fallback for the scenario when router
proxies a request to an application but when the application queries the
content store it returns a redirect type which the frontend app has no
option other than return a 500.