Use unicorn as web server in Procfile
This adds unicorn command to the procfile and changes the
config/unicorn.rb file to inherit from the one defined in
govuk_app_config. For the time being the unicorn config has a call to
worker_processes, however this should be removed once an environment
variable has been set which sets this approriately for environments.