Record the 95th percentile value of publishing delays
Extend the rake task which sends scheduled publishing delay stats to
graphite so that it includes the 95th percentile value as well as the
mean. This will give us an overview of slow requests.

Calculate the value directly rather than using a stats gem because the
available gems weren't very suitable:

- descriptive-statistics
  (https://github.com/jtescher/descriptive-statistics) doesn't calculate
  percentiles in the expected way: it returns nil rather than the
  maximum value if the percentile requested is larger than the data set,
  e.g. the 95th percentile of a data set with less than 20 values.
- SciRuby statsample (https://github.com/sciruby/statsample) is too
  large for what we need: it adds dependencies like a PDF writer and a
  visualisation toolkit.