Log expired publishing intents to Sentry
Log a warning when the housekeeping job cleans up any expired publishing
intents.

This cleanup job used to be necesssary to delete expired intents, but
intents are now deleted when the document is published so there should
no longer be any expired intents.

If this doesn't log any warnings in the next few weeks, it should be
safe to remove the housekeeping task altogether.