Add scheduled publication time to content item response
Add a new field to the API response for a content item: the most recent
scheduled publication date. This will help publishers check whether a
scheduled document was published on time because they can compare the
scheduled time to the actual published time.

This commit reapplies cf17471, which was reverted in 5aef732.