Do not delete publishing intents when content is published
Publishing intents were sometimes being deleted incorrectly. They were
supposed to be deleted when the final version of the document was
published by the whitehall scheduler, but they were sometimes being
deleted slightly early when the "coming soon" version of the document
was republished - this was triggered by the publishing of the document's
attachments. When an attachment is published, dependency resolution in
the publishing-api triggers a publishing of all the linked documents,
including the parent publication.

Deleting the publishing intent early causes a bug, because the
publishing intent is used to set a short cache expiry in the run-up to
the scheduled publishing. If the intent is deleted a few minutes early,
the "coming soon" version is served to users with the default cache
expiry (30 minutes) which delays the appearance of the published
version.

It's safe to leave publishing intents in the DB because they are cleaned
up by the nightly `housekeeping:cleanup_publish_intents` rake task.

This commit ensures that the scheduled publishing log is only updated
once for each scheduled publishing, even though the intent is left in
the DB.

https://trello.com/c/5Pj1ixOa/112-publishing-intent-bug-is-delaying-scheduled-publications