Re-add publishing_scheduled_at field to content item
Add a field to the content item which shows when the current edition was
last scheduled for publication.

This was previously added in commit d1d90f2 and then reverted.
Previously, we added the field at query time, which slowed down the
request by a few ms because of the extra query to
`ScheduledPublishingLogEntry`.

In this commit, we add the new field when saving the content item to the
DB, which means it does not slow down the query.