Remove string interpolation in  where clause
Brakeman complains about string interpolation in SQL statements. In our
case, we were building out a where clause like this:

```
.where("#{kind}s.path" => path
```

This was in fact safe because the method that wrapped our SQL was
only being called from other private methods within the class and none
of these relied on any form of user input. Our code was simply
convenient and slightly DRYer.

Refactor to be more verbose and remove string interpolation and as such,
pass brakeman. Also rename find_content_item_for_field now that we've
moved the query to its calling functions.

Although this code might be slightly more repetitive, it
avoids the need for a brakeman.ignore file in this repository. Arguably
it makes the code slightly easier to reason about as well since
developers don't have to figure out what the string interpolates to.