Don't expose supertypes in content items
We're no longer using supertypes embedded in content items and search
indices, because changing anything requires republishing of all the
affected content. This makes sure that we don't expose the supertypes
in the content item.

The data still is persisted in the database, we'll need to remove the
supertypes from the publishing-api first.

https://trello.com/c/hp5BJD2i