Add Rake tasks for registering routes
In the unlikely event that registering routes with the router failed but
the content-store believes that it succeeded, it will refuse to send the
routes again even if the document is represented. This is because it
caches the routes from the previous item and compares.

These Rake tasks bypass these checks allowing routes to be registered
with the router.