Describe content & publishing API and schemas
As a new joiner, it's quite hard to find all of the different aspects of the
content store in one place. This is an attempt to touch upon all of the key
parts, from reading & writing to the content store, to dealing with different
content types by use of schemas.