Hide 'auth_bypass_ids' field from link expansion
This is private information and should not be shown when accessing
the content store - it is just for internal content store processing.

I've created a 'secret_fields' method which returns a hash, so that
we can easily add more secret fields in future if necessary.

Trello: https://trello.com/c/DR4BMQzv/134-dont-show-authbypassids-in-content-stores-expanded-links