Create auth_bypass_ids field
This creates a new field that is intended to be the destination of data
that is transitioned from the access_limited object. The reason data is
being moved out of the access_limited field is that it creates confusion
given that auth_bypass doesn't do access limiting - quite the opposite
in fact.

To start with this transition this creates a field on ContentItem that
is written to each time data is saved which will be a replica of the
data in access_limited.