Check valid_auth_bypass_ids in nested links
This increases the logic involved in verifying an auth_bypass_id to
consider whether the auth_bypass_id is specified on either the content
item itself or one of it's links.

A user can only be granted access via links if the content itself is not
access limited.

I made a judgement call that auth_bypass_id via links can only apply on
a content items immediate link rather than being nested through a link
of a link. This was mainly on the expectation that was more likely to be
a cause of bugs than expected behaviour.