Content items controller updated for different access requirements
This updates the `can_view?` logic of the content items controller to
reflect the change in requirements for accessing content.

There are now a variety of rules that take effect in the draft stack.

- a user can view content if they are explicitly listed as an authorised
  user or organisation in the access limit of the content;
- a user can view content if they have an auth bypass token that is
  valid for that piece of content;
- a user that meets neither of the above cannot view content that is
  access limited;
- finally, for content with no access restrictions we decide whether a
  user can access based on if they are authenticated

In the live stack content is expected (but not required) to have no
access limiting so users shouldn't be restricted by that. We expect
that in the draft stack authenticating proxy will set auth_user_id or
auth_bypass_id headers, otherwise access won't be granted. So we use
these both being nil as an indication we are on the live stack and
access should be granted.