Private cache headers take precedence over publish intents
This resolves some issues where it was possible for a publish intent to
take priority over the headers of access limited / auth bypassed
content.

This is unlikely to change much as I don't expect publish intents
actually make it to the draft content store, but since they can this
corrects the caching logic.