Resolve text/html content in links
Currently we resolve text/html content in the description and in the
details hash, but not in the links.

This leads to some inconsisteny content items, specifically people where
the links to roles include the different content types but the content
item for the role itself does not.

Example:

```
{
  "api_path": "/api/content/government/ministers/prime-minister",
  "base_path": "/government/ministers/prime-minister",
  "content_id": "845e5811-c0f1-11e4-8223-005056011aef",
  "document_type": "ministerial_role",
  "locale": "en",
  "public_updated_at": "2018-11-05T08:52:39Z",
  "schema_name": "role",
  "title": "Prime Minister",
  "withdrawn": false,
  "details": {
  "body": [
    {
      "content_type": "text/html",
      "content": "The Prime Minister is the leader of Her Majesty's Government and is ultimately responsible for the policy and decisions of the government.\r\n\r\nAs leader of the UK government the Prime Minister also:\r\n\r\n* oversees the [operation of the Civil Service](https://www.gov.uk/government/ministers/minister-for-the-civil-service) and government agencies\r\n* appoints members of the government\r\n* is the principal government figure in the House of Commons\r\n"
    }
  ]
},
```

I'm not sure if there are any frontend apps which currently rely on this
behaviour, although I suspect not since this change hasn't been made
yet. I'm going to deploy this to staging first and see if we notice an
increase in 5xx errors.