Include global data from Homepage in all responses
This adapts the ContentItemPresenter to pull data out of the homepage
(if it exists in as a content item) and embed it in the global field of
all other content items.

This uses a different process to other approaches of embedding shared
data in a content item which is done through the Publishing API approach
of link expansion. This is not seen as a good option for this scenario
because:

1. A change to the homepage which would cause every single GOV.UK page
   to be re-presented to the content store.
2. Any changes made would not appear in content items at the same time
   and would take many hours to propagate.

Given these concerns it seemed a better approach to look up the homepage
automatically. A potential different approach that was considered was to
have a different content item for this data which was special, however
since this item would not have a URL then it couldn't be stored in the
Content Store.

A consequence of this change is that the content store does need a
homepage to exist in order to return the full data, this may cause some
surprises in dev and test environments that may not have published the
homepage.