Enable Lint/UriEscapeUnescape Cop and fix issues
Unfortunately none of the suggestions offered by the Cop work when
trying to fix the issues it raises:

- CGI.escape - this also escapes '/' characters in base paths

- URI.encode_www_form* - this is surprising since we're always trying
to encode a string for a URL, not a form

Instead, I've used 'Addressable::URI.encode', which has the same
behaviour as the original URI.escape. This probably isn't suggested
by the Cop since it's not part of the Ruby standard library.