Add script to try to find possibly unused gems
This script requires simplecov filters to be turned off. At the time
of writing, this only seemed to be easily set in simplecov 0.17.

The script analyses the bundled gems, and makes a comparison with those
found to be used from the specs, as well as a list of gems which we
do not want to flag, such as rails gems.

It outputs all gems which are not covered or are not in the exclusion
list, with the intention that it can help a developer spot unused
gems in the app.