Spike adding /api/content-id/:content_id route
This work explores adding a new public endpoint we can call to
quickly find out what content is represented by a specific content
ID. In absence of a route, the alternative has been to connect to
the VPN, SSH into a Content Store or Publishing API machine and
perform a lookup via the app console.

This idea isn't particularly new. We have a precedent in Whitehall:
https://github.com/alphagov/whitehall/blob/0ef9b6bb99e62e1e7d724b2cd9c7da4233d1bca4/config/routes.rb#L191

This spike works locally. If you app console and compare
`app.get("/api/content")` (the route of the homepage) with
`app.get("/api/content-id/f3bbdec2-0e62-4520-a7fd-6ffd5d36e03a")`
(the content ID of the homepage), you get the same `app.response`.

To make this a viable MVP, we'd need to add an index on the
`content_id`, as currently the lookup is VERY slow (1-2 minutes
per query!). And to make this available via a public endpoint
rather than app console, we'd need to add a 'special route' like
we've done for `/api/content`:
https://github.com/alphagov/special-route-publisher/blob/c47ca1b995315d5abf3c811a321c7520467fe98e/data/special_routes.yaml#L1-L6