Re-raise Router API expectations to ensure 5xx responses
Previously we were treating communication with Router API as a client
error and returning a 422 response to the Publishing API when there was
a Router problem. This is a bit misleading as it is not the case that
updating the client request can resolve the error.

By instead raising an exception these errors will be converted into a
500 response by Rails. This will help us, in the event of more Router
API problems, distinguish server errors from client errors.