Move specialist README subsections into docs/ folder
These subsections repeated information already available in the
docs/ folder, and were too terse to be of much value. I've copied
them over to the most appropriate places in `technical-documentation.md`
and linked to this file in the Further Documentation section, which
is an approach we're commonly beginning to use.