Handle errors for bad URI requests
We occasionally are subject to malicious users trying to discover
weaknesses in our application by sending invalid URIs.

This change updates our handling so that we return a 400 Bad Request
response, rather than recording this as an application error.

I debated about whether to return a 400 (Bad Request) response or a 406
(Not Acceptable) response. However I noticed we already had precedence
for 400 for invalid JSON so I kept consistent with that.