New test for the find specific term rake task
Added as part of CD work to take content-store over the required test coverage.

Also removed the CSV writing from the rake task, as it appeared to have no utility
beyond displaying the data in an appropriate, comma seperated manner, and the presence
of CSV made the task extremely difficult to test. So instead have replaced it with
similar logger output.