Remove schema checking rake task
This task will not have worked for many years as it relies on the format
field which was replaced by document_type and schema in 2016 [1].

From having a look at the Content Store there are only 871 items that
still have a format field so this is likely evidence that this has been
used in at least 6 years.

It is also not tested.

[1]: https://github.com/alphagov/content-store/commit/d135018aca4081ad367e771f00627f82927e36fe