Allow cross-origin requests to access public content endpoint
This addresses https://github.com/alphagov/content-store/issues/1006

This sets a `Access-Control-Allow-Origin: *` header on HTTP responses to
the show endpoint, this will allow requests to this API to originate
(via JS) from hosts other than www.gov.uk.

The reason for adding this is resolve the issue raised on GitHub 1006
[1]. GOV.UK doesn't directly need this as it doesn't make use of client
side requests to the Content Store and, if it did, they'd be from the
same host. However this is added in to reflect that this is indeed a
partially supported public API [2] and that we are not concerned with JS
clients calling it from a different host.

I only put this on the one endpoint as there didn't seem to be any need
to have it on endpoints other than ContentItems#show. I also didn't
implement the HTTP OPTIONS method for the endpoint as I don't think it's
strictly needed and can't see evidence of this enabled on other GOV.UK
cross-origin endpoints:

➜  ~ curl -sI -X OPTIONS https://www.gov.uk/api/search.json\?fields\=publishing_app\&filter_publishing_app\=publisher\&filter_first_published_at\=from:2022-09-01,to:2022-09-30 | grep HTTP
HTTP/2 404

[1]: https://github.com/alphagov/content-store/issues/1006
[2]: https://content-api.publishing.service.gov.uk/