Add a feature flag to disable modifying Router backends on the fly.
Content Store currently updates the backend address in Router every time
a content item or publishing intent is saved. This was originally done
to simplify a long-since-completed migration, but it's now both
unnecessary and causes problems with parallel running for Replatforming,
where the backend addresses necessarily differ between the old and new
installations.

This lets us disable this esoteric behaviour on the Kubernetes side
during the Replatforming pilot phase.