Add note about Jenkins CI dependency
While looking at fixing the Pact verification for Dependabot I noticed
this coupling to Jenkins CI and thought the best thing I could do, save
a large yak shave, was to leave a comment documenting it.

Left unattended, the retirement of Jenkins will mean that the pact
branch is no longer updated and thus will slowly reference an
out-of-date version of Publishing API.

The simplest, but subtly different, way to resolve this would be to
switch to testing against the main branch as with continuous delivery
there is likely only small windows where main is not deployed to
production.