Remove placeholder logic
Placeholder schemas were added to hold documents that were rendered by
Whitehall and not yet ready to be presented to end users.

This added some complexity to Publishing API and Content Store, as these
routes did not need to be registered with Router.

Therefore removing this logic now all rendering has been migrated out of
Whitehall.