Add test for gone content item with nil redirects
The `GonePresenter` in Publishing API does not include a value for
`redirects` [1], therefore it will be `nil`.

Our current test for 'gone' content items assumes the value of
`redirects` will be `[]`, which is not true.

This appears to be causing an error when we try to unpublish former
redirects with a 'gone' content item in their place.

Therefore adding a test that reproduces the problem we have seen in
production.

1: https://github.com/alphagov/publishing-api/blob/712b548ee6e75c85a15d808d667678c68e2d3145/app/presenters/gone_presenter.rb#L42-L56