Remove requires for `update_lock`
This is a model so is automatically loaded by Rails. We don't need to
explicitly include it anymore.

As recommended by Rails, the cache format version has been fixed at 7.0
to allow us to rollback. This line can be removed when we are certain
that Rails 7.1 is running correctly in production.