Remove unnecessary flaky assertion
This test (which simply makes a GET request on the content item
endpoint) was testing whether the returned content item included a
`public_updated_at` time that was within one second of now.

This relied on the record being created within 1 second of the GET
request being made by the test, therefore sometimes failing.

However we are not testing the creation of the record here, so it is
not required to check the timestamp.