Update Rake task logging expectation
a8b0f53 bumped `govuk_app_config` from 9.11.2 to 9.12.0, which set the
`time_zone` to London in all GOV.UK apps. This appears to change how the
timezone is reported when calling `updated_at` on a record: what was
`UTC` is now `+0000`. This change resulted in a couple of failing tests,
but only related to console logging within a Rake task, so we should be
fine just to update our expectations here