Notify Router only on specific column updates
Currently, we send a notification to the Router to update route tables
whenever there is any change to the content_items (or publish_intents)
tables. However, the Router only needs to update the route table when
changes affect specific columns that determine the routes.

This change should reduce unnecessary reloading.