Add data migration to remove `info-frontend` rendered content items
We have some content items in the database which are marked as being
rendered by `info-frontend`. This application has not existed for a
number of years and the content items result in a 404 error when they
are viewed by the public.

They have already been removed from Publishing API (all the records
deleted, instead of being unpublished), so adding a migration to clean
up the Content Store database to no longer include them.